# feat: improve perf of groupMessagesForPartition (~600x in production use-case)